requirements_capstone.txt
# Core
numpy==1.26.4
pandas==2.2.2
matplotlib==3.7.1
scikit-learn==1.2.2
seaborn==0.12.2

# Deep Learning (TensorFlow/Keras)
tensorflow==2.15.0

# Common TF dependency compatibility on Windows
protobuf<5

# Optional but useful
pydot
tqdm

# Jupyter kernel support
ipykernel

# Web app
streamlit
pillow
